+++ packages/gasket-cli/test/unit/hooks/init.test.js
@@ -45,6 +45,7 @@ describe('init hook', () => {
     delete process.env.GASKET_ENV;
     delete process.env.GASKET_CONFIG;
     delete process.env.GASKET_ROOT;
+    delete process.env.GASKET_COMMAND;
   });
 
   it('ends early for create command', async () => {
@@ -61,12 +62,14 @@ describe('init hook', () => {
     assume(process.env).not.property('GASKET_ENV');
     assume(process.env).not.property('GASKET_ROOT');
     assume(process.env).not.property('GASKET_CONFIG');
+    assume(process.env).not.property('GASKET_COMMAND');
 
     await initHook({ id: 'build', argv: [] });
 
     assume(process.env.GASKET_ENV).equals('development');
     assume(process.env.GASKET_ROOT).equals('/path/to/app');
     assume(process.env.GASKET_CONFIG).equals('gasket.config');
+    assume(process.env.GASKET_COMMAND).equals('build');
   });
 
   it('gets the gasket.config', async () => {