+++ packages/gasket-cli/src/config/default-plugins.js
@@ -2,5 +2,6 @@ module.exports = [
   require('@gasket/plugin-command'),
   require('@gasket/plugin-lifecycle'),
   require('@gasket/plugin-metadata'),
-  require('@gasket/plugin-start')
+  require('@gasket/plugin-start'),
+  require('@gasket/plugin-git')
 ];