+++ packages/gasket-cli/CHANGELOG.md
@@ -1,5 +1,6 @@
 # `@gasket/cli`
 
+- Add git plugin to CLI default plugins
 - Address createContext helper types
 - expose `GASKET_COMMAND` on process.env